Move UserProfile to `net.mamoe.mirai.data`
